Fixing class templates
git-svn-id: http://my-svn.assembla.com/svn/aion-emu/trunk@454 2af79824-a935-44bf-bf41-fd8ed1897d0c

+++ AE_GameServer/src/com/aionemu/gameserver/model/templates/StatsTemplate.java
@@ -59,14 +59,14 @@ public class StatsTemplate
 	private int will;
 
 	@XmlElement(name = "walk_speed", required = true)
-	private int walkSpeed;
+	private float walkSpeed;
 	@XmlElement(name = "run_speed", required = true)
-	private int runSpeed;
+	private float runSpeed;
 	@XmlElement(name = "fly_speed", required = true)
-	private int flySpeed;
+	private float flySpeed;
 
 	@XmlElement(name = "attack_speed", required = true)
-	private int attackSpeed;
+	private float attackSpeed;
 
 	@XmlElement(name = "evasion", required = true)
 	private int evasion;
@@ -131,22 +131,22 @@ public class StatsTemplate
 
 	/* ======================================= */
 
-	public int getWalkSpeed()
+	public float getWalkSpeed()
 	{
 		return walkSpeed;
 	}
 
-	public int getRunSpeed()
+	public float getRunSpeed()
 	{
 		return runSpeed;
 	}
 
-	public int getFlySpeed()
+	public float getFlySpeed()
 	{
 		return flySpeed;
 	}
 
-	public int getAttackSpeed()
+	public float getAttackSpeed()
 	{
 		return attackSpeed;
 	}
